+++ js/data.js
@@ -1,0 +1,5 @@
+const names = ['Victor', 'Anna', 'Petr', 'Vasiliy','Nadezhda','Irina','Vadim','Olga','Maxim','Arina','Alexey'];
+const comments = ['Всё отлично!', 'В целом всё неплохо. Но не всё.', 'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.','Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.','Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.','Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+const description =['Недостаточно овладеть премудростью, нужно также уметь пользоваться ею', 'Тем, кто хочет учиться, часто вредит авторитет тех, кто учит.', 'Где умирает надежда, там возникает пустота.', 'Подумай, как трудно изменить себя самого, и ты поймешь, сколь ничтожны твои возможности изменить других.', 'Невежество — ночь ума, ночь безлунная и беззвездная.', 'Есть только одно благо — знание, и только одно зло — невежество.'];
+
+export {names, comments, description};
